rename to yjs-connection-manager
